added tanh() before output layer
Since the state will keep accumulating as time goes on (because at first the remember gates are biased to remember), the outputs from the output DenseLayer were probably very large and saturating the SoftmaxLayer. To fix this, I added a tanh() to squish the current state before feeding it to the output layer.